Fixes rows not reloading after changing the strategy
The problem is the so called "reducer" which is nothing more than a
function that takes current component's state and returns a new state by
reacting to a given action (just a switch statement to select the action
that was requested) in the case of this application is a new kind of
thing that I invented to solve a seemingly difficult problem (the
difficulty being making it work the React Way!) is a dynamic object in
which these action selectors that I mention are also dynamic and have an
id (this is part of what makes it slow too).

Well they need to be injected into the so called redux store when we
change the strategy, but there's no guarantee that it will be injected
before initialization so that it's ready when getSnapshot() is called.

The reason why it works sometimes is because sometimes, it makes it into
the store before the backend has responded (my connection is slower than
yours).But I never see the spinners because when that happens we
certainly don't have a reducer yet!!!