Several fixes and new stuff
Mainly handling client errors is working, as well as a new timer that
waits for sumbit or pricing requests which need in turn to wait for
SignalR messages to hide the waiting windows.

This timer will remove the waiting windows and show and alternative
error window that is non-obstrusive and suggests to the user to retry or
refresh the window.